docker_setup/requirements.txt
boto3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==4.2.14
django-storages
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn
pillow
psycopg2-binary
python-decouple==3.8
pytz
